componentise reused LabelAndText couples
